requirements.txt

fastapi==0.115.7
uvicorn[standard]==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.14.0
pyjwt==2.10.1
bcrypt==4.2.1
python-multipart==0.0.12
websockets==14.1
asyncpg==0.30.0
torch==2.5.1
numpy==2.2.1
opencv-python==4.10.0.84
pyopencl==2024.2.7
